fix: posthog insight tests depend on ee code
